expences_proj/requirements.txt
asgiref==3.2.10
confusable-homoglyphs==3.2.0
Django==3.1.2
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-registration==3.1.1
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
